requirements.txt
google-cloud-speech==2.16.1
google-cloud-storage==2.10.0
pyaudio
numpy
tk
threading
queue
sounddevice
soundfile
python-dotenv